pyproject.toml
[project]
name = "maxibitsolutionstestproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "python-telegram-bot (>=22.0,<23.0)",
    "opentelemetry-proto (>=1.33.0,<2.0.0)",
    "apache-airflow[celery] (>=3.0.1,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "flake8 (>=7.2.0,<8.0.0)"
]

[tool.poetry]
packages = [{include = "maxibitsolutionstestproject", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
